fix(build): hook 改用 BaseFeature 内置 hookBefore 管理（registerUnhook 自动卸载），修复 += 收集 Unit 导致编译失败；TitleTagCommon setStroke 恢复 Int width
- BaseFeature.hookBefore 内部 registerUnhook 返回 Unit，不能收集 HookHandle；改用成员扩展 onClick.hookBefore(100){}，由 BaseFeature.unhooks 统一管理，disable/unhookAll 自动清理
- unhookAllSendHooks/sendHookHandles 删除，reinstall 直接 unhookAll()+installSendHooks()
- GradientDrawable.setStroke(width: Int, color: Int) 需要 Int，恢复 .toInt()（仅 cornerRadius/shadowLayer 需要 Float）

+++ app/src/main/java/dev/ujhhgtg/wekit/features/items/chat/TitleTagCommon.kt
@@ -154,7 +154,7 @@ internal fun applyTitleStyle(view: TextView, entry: TitleEntry) {
         intArrayOf(bgStart, bgEnd),
     )
     bg.setCornerRadius(9f * density + 0.5f)
-    bg.setStroke(1f * density + 0.5f, Color.argb(180, 255, 239, 190))
+    bg.setStroke((1f * density + 0.5f).toInt(), Color.argb(180, 255, 239, 190))
     view.background = bg
     view.setTextColor(fg)
     view.textSize = 10f
